cogy: Make the request a POST
Also, the whole Cog environment from the Relay is passed as-is to the
backend app.

See https://cog-book.operable.io/#_command_environment_variables for
more on what is exactly passed.

This is a breaking change and works only with Cogy gem v0.3.0 or later.

Part of https://github.com/skroutz/cogy/issues/25